create tode list and delete
