+++ src/App.js
@@ -1,45 +1,28 @@
 import React, { Component } from 'react';
-// import Dataapp from './Dataapp';
-import Nestapp from './Nestapp'
-import AddNiga from './AddNiga';
+import Todos from './Todos';
 
 class App extends Component{
-
   state = {
-    ningas:[
-      { name: 'Ryu',  age: 30, belt: 'black', id: 1},
-      { name: 'yoshi', age: 40, belt: 'green', id: 2},
-      { name: 'Crystal', age: 25, belt: 'pink', id: 3}
+    todos:[
+      {id:1, content: 'Buy same goods'},
+      {id:2, content: 'Play cards'}
+  ]
+}
 
-    ]
-  }
-
-// function to add data
-  AddData = (savedata) =>{
-    savedata.id = Math.random();
-    let ningas = [...this.state.ningas, savedata];
-    this.setState({
-      ningas: ningas
-    })
-
-  }
-
-  // function for deleting data
-  deleteData =  (id) => {
-    let Datadelete = this.state.ningas.filter(savedata =>{
-      return savedata.id !== id
-    });
-    this.setState({
-      ningas: Datadelete
-    })
-  }
+// function to delete todos.
+deleteTodo = (id) => {
+  const todos = this.state.todos.filter(todo => {
+    return todo.id !==id
+  });
+  this.setState({
+    todos
+  })
+}
   render(){
     return (
-      <div className="app-connect">
-        {/* <Dataapp /> */}
-  
-        <Nestapp deleteData={this.deleteData} ningas = {this.state.ningas} />
-        <AddNiga AddData = {this.AddData} />
+      <div className="app-connect container">
+        <h1 className="center blue-text">Todo's</h1>
+        <Todos todos={this.state.todos} deleteTodo={this.deleteTodo} />
       </div>
 
     )
